requirements.txt
pandas==2.2.0
openpyxl==3.1.2
pdfplumber==0.10.3
pytesseract==0.3.10
rapidfuzz==3.6.1
unidecode==1.3.8
Pillow==10.2.0